refactor: rename hook file
