+++ apps/calendar/src/hooks/timeGrid/useTimeGridScrollSync.ts
